pk-client: Fix a possible use-after-free under pk_client_cancel_cb()
When cancelling a transaction, even if it fails, the associated transaction
still receives its async callback, thus also frees the 'state' structure.
That means, when the cancel fails, it should not free the 'state'.

This can be reproduced by a race condition between cancelling the operation
by the caller and delivery of the 'Cancel' D-Bus call. The call order is:
- the caller cancels the 'state->cancellable_client' cancellable
- the PkClient executed the 'Cancel' D-Bus method for this operation
- before the 'Cancel' D-Bus call is finished the operation finishes,
  freeing the 'state' as well
- the 'Cancel' D-Bus method call is finished, possibly with an error like
  "Transaction is already finished". As the 'state' is already freed, trying
  to free it causes a use-after-free error and an application crash.

Signed-off-by: Richard Hughes <richard@hughsie.com>