dnf-backend: honor install_weak_deps=False if it is there
Currently, even if we have "install_weak_deps=False" in
/etc/dnf/dnf.conf, `pkcon install` still tries to install all
the recommended packages.

For avoiding that, we need to inform libdnf that we will
solve the goal of the transaction ourselves (by means of the
dnf_transaction_set_dont_solve_goal() API) and then explicitly
set the flag for ignoring the weak dependencies.

This fixes issue #486 and also solve
https://bugzilla.redhat.com/show_bug.cgi?id=1955484