PkClient: Improve thread safety on an operation cancellation
The PkClient operates in two stages. The first is an "active" stage,
which makes one or more D-Bus calls. The second is a "passive" stage,
when it is waiting for the "Finished" D-Bus signal from the daemon.
The cancellation can come in any time. This change keeps the cancellation
up to the D-Bus calls during the first stage and actively stops the waiting
during the second stage.

As the GCancellable is cancelled in a different thread, and any D-Bus call
is also done and its result delivered in that thread default main context,
rather than in the main context when the PkClientState had been created,
the PkClientState is switched to a GObject descendant, to be able to benefit
from the GWeakRef, which can avoid use-after-free easily.

The GDBusProxy callbacks can be schedules in the middle of the object free,
thus the added callbacks also benefit from the GWeakRef usage.

All of this helps to better cover the PackageKit daemon disappear during
an ongoing call to it.

Closes https://github.com/PackageKit/PackageKit/issues/489