aptcc: Propagate install mode info in all emitted package-IDs
With this change, information about the installation mode of a package
(manual or automatic) is added to package-IDs emitted by APTcc. We will
always emit this information for installed packages, while only emitting
it in case an installation is requested for available packages to
preserve backwards compatibility.
To indicate that the package does not yet hold the desired state, the
state-change is indicated with a plus sign.

This will enable a future change to make updates behave smarter and not
loose the package install mode when an update installs new packages.
It's a hack, but the best solution to this problem that I see currently.
Helps: #450