trivial: aptcc: Make state-change-check more robust
A plus sign will be filtered out already, but in case someone did manage
to sneak it in as a repository origin name, and some other issue
resulted in the package-IDs not being checked against the local cache
(and "our" prefixing was skipped), an attacker could change the package
states to auto and brick a system.
This is an extremely unrealistic scenario, but still, we should make
this check complete and include the colon-separator, as that will never
be permitted as repo origin and is sanitized for at every occasion.