aptcc: Display an excerpt of dpkg's log on error instead of not failing at all
This is useful when e.g. a dpkg hook script fails. Previously, we would
just pretend the package was installed, even if, in fact, nothing
happened. Just emitting a "something bad happened!" error is also wrong,
so instead we will show the latest messages of dpkg's log, which will
usually be concise and also give a pretty good hint as to what went
wrong.

Resolves: #281