pk-transaction: Add Packages signal
This has reduced the number of signals emitted during starting up
gnome-software from 5711 `Package` signals, to 66 `Packages` signals
(one per transaction).

The signal is used only if the client sets the
`supports-plural-signals=true` hint on a transaction. Otherwise the old
`Package` signal will be used. This means the new signal is
backwards-compatible.

If the hint is set, the new signal will be used in parallel with the old
one, with some operations or backends using the old signal, and some
using the new one. The content of the signals will not be duplicated
(i.e. if `Packages` is emitted for N packages, there will not also be N
`Package` signal emissions in parallel).

The new signal is currently only implemented for the dnf backend. Other
backends can implement it at their leisure.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>