pk-transaction: Add UpdateDetails signal
Like the `Packages` signal added a few commits previously, the
`UpdateDetails` signal is a plural version of the existing
`UpdateDetail` signal.

This has reduced the number of signals emitted during starting up
gnome-software from 632 `UpdateDetail` signals (for a corresponding
number of pending dnf updates) to 1 `UpdateDetails` signal.

The new signal is only used if the client sets the
`supports-plural-signals=true` hint on a transaction. Otherwise the old
`UpdateDetail` signal will be used. This means the new signal is
backwards-compatible.

If the hint is set, the new signal will be used in parallel with the old
one, with some operations or backends using the old signal, and some
using the new one. The content of the signals will not be duplicated
(i.e. if `UpdateDetails` is emitted for N updates, there will not also
be N `UpdateDetail` signal emissions in parallel).

The new signal is currently only implemented for the dnf backend. Other
backends can implement it at their leisure.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>