zypp: restore pool status after simulating an update
Previously, when simulating an update, doUpgrade() or doUpdate() is
called. But the changes to the pool never got commited nor cancelled.
This leads to the result that locked packages are updated in a later
update/upgrade.

Fix that by restoring the ResPool status to the state before simulating
an update.

https://bugzilla.suse.com/show_bug.cgi?id=1199895