trivial: aptcc: copy a string directly to std::string, which is expected below
Below, each of the many invocations of starts_with(str, ...) takes
std::string as the argument, so str would anyway be converted to
std::string each time (and even copied again each time).

Improves: 323cc7483 trivial: aptcc: Some trivial cosmetic changes

(What really caused me to do this change was a non-GCC compiler having
troubles with that implicit conversion after the type of str was
annotated with g_autofree, which in its turn we had to change to
g_autofree_edg as another workaround. But nevertheless, this change can
be justified on its own, as at the top of this commit message.)