Store sender command-line in transaction database again
Previously the client command-line was never read and therefore not
added to the transaction database, as the transaction role was not yet
set when a sender was provided early on (so, the role was always UNKNOWN
and pk_dbus_get_cmdline() was hence not called).

The command-line role dependency was added as a performance
optimization, so the new code tries to be as performant as possible, and
no measurable difference to the previois code was observed.

If we do need to get the sender PID and UID, we use the "new"
GetConnectionCredentials D-Bus call, which improves performance over
making two calls to get PID and UID instead.