pk-client: Signal a transaction as finished if it disappears off the bus
The client code already does this if the `Finished` signal is received,
or if the packagekitd process completely disappears off the bus.
However, if the transaction is destroyed by the daemon without emitting
`Finished`, it’s possible for the client proxy for it to never complete.

Fix that by calling `pk_client_state_finish()` in a few additional
locations.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>