pk-transaction: Check that Finished signal is emitted at least once
It feels like transactions should always emit `Finished` before being
removed from the bus, either signalling success or containing error
details. Emitting a signal like this allows client code to track
transactions properly.

A lot of the code paths in `PkTransaction` ensure that `Finished` is
emitted. However, there is at least one code path where it’s not, and
that was causing the assertion in this commit to fail:
```
4  0x00007ffff77408ae in g_assertion_message_expr () at /lib64/libglib-2.0.so.0
5  0x000000000044f9e7 in pk_transaction_dispose (object=0x62500010af80)
   at ../../source/PackageKit/src/pk-transaction.c:5512
6  0x00007ffff7eeb8d4 in g_object_unref () at /lib64/libgobject-2.0.so.0
7  0x00000000004acb4b in pk_scheduler_item_free (item=0x607000311210)
   at ../../source/PackageKit/src/pk-scheduler.c:201
8  0x00000000004ad140 in pk_scheduler_remove_internal
   (scheduler=0x61d000004120, item=0x607000311210)
   at ../../source/PackageKit/src/pk-scheduler.c:233
9  0x00000000004b0083 in pk_scheduler_no_commit_cb (user_data=0x607000311210)
   at ../../source/PackageKit/src/pk-scheduler.c:590
```

That’s the code path to handle a transaction which is created, but which
then isn’t used by the client before a timeout triggers.

The documentation isn’t clear on when `Finished` should be emitted.
There’s also the `Destroy` signal which *is* always emitted before the
transaction is removed from the bus.

On the basis that every other code path emits `Finished`, this commit
also makes the `pk_scheduler_no_commit_cb()` path emit `Finished` too.

Now clients can use `Finished` and bus name notifications to reliably track
a transaction’s lifetime.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>