pk-client: Fix use-before-set of PkProgress on cancellation
If `pk_client_state_new()` is called with a `GCancellable` which is
already cancelled, it immediately finishes itself with a cancellation
error. That’s correct, but the code wasn’t handling the fact that this
happens before `client->progress` is initialised, so was assuming it was
non-`NULL`.

See https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2331

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>