pk-client: Move cancellable disconnect to dispose
This avoids the potential for `g_cancellable_disconnect()` to be called
from within `pk_client_cancellable_cancel_cb()` (via
`pk_client_state_finish()`). This would result in a deadlock, as
`g_cancellable_disconnect()` is not allowed to be called from within
`g_cancellable_cancel()`.

This change should be safe in terms of the cancellation handler being
called twice, as it clears `state->res` on cancellation, and bails out of
`pk_client_state_finish()` if that’s unset on entry the next time.

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>

Fixes: https://github.com/PackageKit/PackageKit/issues/665