dnf/zypp: Add support for language(ll_CC) to what-provides queries
This was already documented in the [‘package component provides naming’
spec](./docs/provides-component-naming.txt), so should be
uncontroversial.

This is part of a plan to drop the fedora-langpacks plugin from
gnome-software, and replace it with the equivalent of a call to `pkcon
what-provides 'language(zh_CN)'` (for example).

See https://gitlab.gnome.org/GNOME/gnome-software/-/merge_requests/2055#note_2134286

Signed-off-by: Philip Withnall <pwithnall@gnome.org>