pk-client: Drop strong ref on PkClient in PkClientState
From some quick looking around, it appears that there are one or more
`PkClientState`s per `PkClient`, and that the `PkClient` has a longer
lifetime. The `PkClientState`s are all tracked in `PkClient->calls`.

So it makes sense for `PkClient` to hold the strong ref on the
`PkClientState`, and the `PkClientState` to only hold an unreffed
pointer on the `PkClient`, so that the refs match the expected
lifetimes.

If both types were to hold strong refs, there’d be a refcount cycle.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #756