pk-client: Avoid array being modified during iteration
Cancellation of a `GCancellable` is synchronous, so this iteration over
a load of cancellables could result in their callbacks modifying the
array (via `pk_client_state_remove()`) while iteration is ongoing,
causing elements to be skipped.

Avoid that by iterating over a temporary copy of the array. Use a
relatively new GLib function for this (`g_ptr_array_new_from_array()`) —
this requires a dependency version bump. If that’s problematic, the new
function can be replaced with some open coding instead.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Helps: #756