Fix check in pk_engine_proxy_logind_cb
The original code does not produce a warning on a system without the
org.freedesktop.login1 service. The proxy object gets created neverthless and
this glib issue [1] hints that it is the owner name that should be checked
for NULL.

The correct fix belongs to glib but for now workaround it on the PackageKit side.

[1] https://gitlab.gnome.org/GNOME/glib/-/issues/879