pk-client-helper: Do not crash on edge-case GIO error
The GIOChannel API provides both a `status` and an `error` output.
The status can be NORMAL, AGAIN, EOF, ERROR.

It is not the case, as the code previously assumed, that
(status != NORMAL) => (error != NULL).
It is unfortunately not even true that
(status == ERROR)  => (error != NULL).

We could try handling AGAIN and EOF better, but for now let's avoid
crashing with a NULL pointer dereference.