Allow changing offline update action without polkit authentication
PackageKit does not offer an API to change the post-update action between
reboot and shutdown, so when KDE Plasma wanted to expose this feature it got
implemented by calling Trigger(action) at logout time [1] [2].

That seems to work fine for offline-updates because by default Trigger() can be
called without polkit asking for authentication.
However, for system-upgrades instead TriggerUpgrade() requires authentication.
That causes a polkit dialog to appear behind the logout screen, causing bug [3].

If a system-upgrade was already triggered, merely changing the post-update
action should not be a sensitive operation so we can allow that to proceed
without prompting polkit.

[1]: https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/4649
[2]: https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/4596
[3]: https://bugs.kde.org/show_bug.cgi?id=495538