apt: Implement enable/disable support for deb822-style repo sources
In the long run, this will require a complete rewrite of the truly awful
SourcesList editor class, but in the meanwhile adding this support will
at least allow for non-destructive enabling and disabling of software
sources via PackageKit no matter which style they are saved in.

For the fairly inevitable rewrite of all this code, we can hopefully do
that at a time when we can drop support for the legacy sources.list
format.

Resolves: #763