pk-client: Use fast array removal
We don’t care what order the `calls` array is in, so might as well use
the ‘fast’ array remove method to speed things up very slightly.

It moves the final element of the array in place of the removed one,
rather than shifting all the elements after the removed one down by one
place (which is a `memmove()` call).

Signed-off-by: Philip Withnall <pwithnall@gnome.org>