pk-client: Tidy up waiting_for_finished implementation
The effective purpose of this boolean is to indicate that the
`PkClientState` is now waiting for a signal before it can be finished
(i.e. marked as successful or an error, and its `g_task_return_*()`
method called).

This is in contrast to earlier in the `PkClientState`’s lifecycle, where
it was waiting for a sequence of D-Bus method call replies. Because of
the way GLib asynchronous calls work, a reply was always guaranteed for
those. A signal is not always guaranteed, so this boolean sort of
signifies that we’re in that slightly dubious period.

Tidy up the comments to list all the signals which would result in the
`PkClientState` being finished, and ensure the code in all three of them
is consistent.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>