pk-client: Drop double-ref call
I believe the intention of this `g_object_ref()` was to keep the
`PkClientState` alive while waiting for one of the completion signals
(`Finished`, `Destroy`, or `NameOwnerChanged`). I’m not 100% sure
because there was no documentation or code comment, and I couldn’t find
an explicitly paired `g_object_unref()` call.

In any case, from my testing, and coupled with the fact that since a few
commits ago we now explicitly hold a strong reference on the
`PkClientState` in the `->calls` array (which it is a member of while
waiting for any of these signals), I think this `g_object_ref()` is now
a double-ref and should be dropped.

While testing this change, I haven’t seen any use-after-free problems
which would indicate that it’s a bad change. And I have seen fewer
leaks.

I would love to be more certain that this change is correct, but
unfortunately the architecture behind this code has been lost in the
mists of time, and despite doing several hours of archaeology while
putting together this branch, I’m still no closer to understanding it.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>