pk-client: Make pk_client_get_progress_state_finish() match rest of code
`pk_client_get_progress_state_finish()` is used as an analogue of
`pk_client_state_finish()`, but for `pk_client_get_progress_async()`. As
the commit which introduced the code implies, this needs to be slightly
different from the rest of the methods in `PkClient` because it’s meant
to query the status of a transaction without adopting it.

However, it still uses the same `PkClientState` mechanisms as the rest
of the file and, in particular, has a new `PkClientState` to represent
its asynchronous context. This `PkClientState` needs to be tracked and
freed just like any other.

So, make `pk_client_get_progress_state_finish()` mirror
`pk_client_state_finish()` in how it handles the `PkClientState`
lifecycle.

The `GCancellable` handling code which is being removed is already
present in `pk_client_state_dispose()`, so defer to using that.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>