pk-client: Track PkClientState from earlier in its lifecycle
Call `pk_client_state_add()` on a `PkClientState` as soon as it’s
constructed. This adds it to the `->calls` array in the `PkClient`.

Previously, the `pk_client_state_add()` was only called after
`pk_control_get_tid_async()` and `g_dbus_proxy_new_for_bus()` had
succeeded for the operation, just before calling the actual transaction
method the caller is interested in (e.g. `SearchDetails()`).

In particular, this means that if `pk_client_cancel_all_dbus_methods()`
is called while a call to `pk_control_get_tid_async()` is in flight,
that call will now be successfully cancelled. Previously it would not.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>