pk-client: Fix an assertion failure in PkClient
This was introduced by one of my earlier commits, but I think it makes
sense to fix in a separate commit so the context can be documented.

It turns out that `state->res` (which is a `GTask`) can hold the last
reference to the `PkClient`. If we clear `state->res` before calling
`pk_client_state_remove()`, then `pk_client_finalize()` gets called as a
result of clearing `state->res`, and its assertion that `->calls` is
empty fails, because the `pk_client_state_remove()` call to remove the
final `PkClientState` from it has not quite yet been called.

So, revert to calling `pk_client_state_remove()` before clearing
`state->res`, but hold a new strong ref over the two calls to avoid the
possibility of `pk_client_state_remove()` dropping the final ref on the
`PkClientState` and leading to a use-after-free when we try to clear
`state->res`. Whew.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>