backends/eopkg: Fix atrocious get_updates() performance
_extract_update_details() is parsing an XML file that is well over
100k of lines, for every single update!!! This is beyond stupid,
especially when it directly bypasses the pickle cache.

The only information we needed to get was whether the package was
a security update or not, bugfix matching was not implemented.

Luckily you can get the history from the packagedb pickle cache instead.
Copy and paste some lines from solus-sc to do what we want.

Before: 1m10.858s | Now: 0m0.573s for 163 available updates.