backends/eopkg: Overhaul SimplePisiHandler to set percentage updates
Listen to pisi callbacks in order to calculate the global percentage,
in a short amount of words; getting to this place wasn't easy as the
callback system pisi uses isn't conductive for a global progress bar.

In part, any forced removals such as deprecations or replaces will reset
the packagestogo counter and run usysconf triggers before applying
any upgrades for example. As we calculate the global percentage off the
packagestogo callback this messes things up considerably.

This handler is used by any "privileged" operations such as upgrade,
remove, etc.