backends/eopkg: Fix resolve() to handle multiple values
Previously, it would only parse the first value.

Additionally, if the value does not exist, do not exist with an error,
and, continue to parse the remaining values to comply with the spec.
This is important for software centers which will merge appstream data
from multiple catalogs (e.g. distro, flathub, etc) then resolve all the
package_ids it finds.

With this commit, software centers should now be able to parse
packages from this backend reliably.