client: Send SIGTERM to debconf helper instead of SIGQUIT
Sending SIGQUIT is weird, because we do not want to dump core. This is
not a failure scenario.
Changing this will also allow us to simplify the Qt debconf helper a
little in future.

We may decide to be cleaner and send SIGKILL after a delay, if we
encounter misbehaving helpers to be a problem.