policy: Allow inactive users to refresh sources
In remote-access scenarios (e.g. with xrdp) users will otherwise get
spammed with password-prompt dialogues from applications wanting to
refresh update information.

There is no real attack vector here that would justify this restriction
for nonlocal/inactive users, except for a DoS attack, which can be
achieved in many other ways easier (and will even be hard with this as
backends complete immediately when the cache is up to date).