pkcon: Don't filter out removed packages from the update target set
This was introduced in b10185b72b9eb16f24be13e6712c06e7943011f8 but it
papers over an older issue in the APT backend and we should rather fix
that properly.

All updates (including to-be-removed package-ids) should be submitted to
the backend so it can do the right thing.

Having this filter also breaks more complex updates on Debian when using
pkcon, as it removes freedom for the solver (barring it from removing
packages).