apt: Fully resolve update candidates, allowing removals
This should allow complex upgrades to succeed and should be resilient
against bad input. It also allows passing the output of GetUpdates()
back into the backend.

The drawback is that we resolve updates twice, which is a bit wasteful -
but I see no better way to do this. We can not trust the frontent to
mess things up afterall, so we need to re-resolve.