offline-updates: Assume plymouth>=0.9.5 by default
If Plymouth isn't in the build environment at compile-time, we build for
an older version of Plymouth, degrading the experience. With our
required Plymouth version being over 5 years old, we can assume it to be
present everywhere at this point and default to it instead.

People who for some reason haven't upgraded Plymouth but are using a
newer PackageKit will get an error in the journal and hopefully will
then just upgrade the dependency.