pkgctl: Add 'monitor' subcommand
This command is special, since it does not reply on the machinery the
pkgctl context establishes, so we had to use some pretty ugly hacks to
make it work in the same tool...

Ultimately, this is not too bad though, and it is nice to have all
commands in the same command-line utility.
In the long run, "pkgctl monitor" will replace pkmon.