freebsd backend: Feed package names without versions to the update command
I falsely assumed that the update command accepts a list of installed packages
(which have old versions). According to #913 it should update the specified
packages to their latest versions. In FreeBSD package we can't have multiple
versions of the same package installed, so there is no sense in passing the
version part of the package name to 'update'.