pk-client: Fix assertion failure when PkClientState is cancelled early
If the `GCancellable` passed to `pk_client_state_new()` is already
cancelled at the time of the call, the cancellation callback
(`pk_client_cancellable_cancel_cb()`) would be called synchronously and
remove the newly created `PkClientState` from the `PkClient.calls`
array.

Unfortunately, this happened immediately before the `PkClientState` was
added to `PkClient.calls` for the first time, in the remainder of
`pk_client_state_new()`.

This meant that the now-defunct `PkClientState` would hang around in
`PkClient.calls` forever, until it triggered an assertion failure in
`pk_client_finalize()` due to `PkClient.calls` not being empty.

Fix that by adding the `PkClientState` to `PkClient.calls` _before_
connecting its cancellable. The code has to be safe to call this way
round, as the cancellation callback could theoretically be called at any
time after the cancellable is connected.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: https://gitlab.gnome.org/GNOME/gnome-software/-/issues/2873